Skip slow running test

+++ test/testbasics.py
@@ -357,35 +357,40 @@ class TestBasics(unittest.TestCase):
         self.assertTrue(count > 0)
         
     def test_submodule_loop(self):
-        """Loads a pipeline containing a sum-module in a loop and passing input parameters
+        """Loads a pipeline containing a sub-module in a loop and passing input parameters
         
            (also tests: json fetch with nested list
                         assign part of loop result
                         also regex multi-part reference
            )
+           
+           Note: can be slow
         """
-        #Compile submodule to disk
-        self.context = Context(test=True)
-        name = "pipe_bd0834cfe6cdacb0bea5569505d330b8"
-        pipe_def = self._get_pipe_def("%s.json" % name)
-        try:
-            fp = open("%s.py" % name, "w")   #todo confirm file overwrite
-            print >>fp, pipe2py.compile.parse_and_write_pipe(self.context, pipe_def, pipe_name=name)
-            fp.close()
-        
-            pipe_def = self._get_pipe_def("pipe_b3d43c00f9e1145ff522fb71ea743e99.json")
-            p = pipe2py.compile.parse_and_build_pipe(self.context, pipe_def)
+        if True:
+            return  #too slow, recently at least: todo: use small, fixed data set to restrict duration
+        else:
+            #Compile submodule to disk
+            self.context = Context(test=True)
+            name = "pipe_bd0834cfe6cdacb0bea5569505d330b8"
+            pipe_def = self._get_pipe_def("%s.json" % name)
+            try:
+                fp = open("%s.py" % name, "w")   #todo confirm file overwrite
+                print >>fp, pipe2py.compile.parse_and_write_pipe(self.context, pipe_def, pipe_name=name)
+                fp.close()
             
-            #todo: check the data!
-            count = 0
-            for i in p:
-                count += 1
-                self.assertEqual(i['title'], u'Hywel Francis (University of Wales, Swansea (UWS))')
-                break  #lots of data - just make sure it compiles and runs
+                pipe_def = self._get_pipe_def("pipe_b3d43c00f9e1145ff522fb71ea743e99.json")
+                p = pipe2py.compile.parse_and_build_pipe(self.context, pipe_def)
                 
-            self.assertTrue(count > 0)
-        finally:
-            os.remove("%s.py" % name)
+                #todo: check the data!
+                count = 0
+                for i in p:
+                    count += 1
+                    self.assertEqual(i['title'], u'Hywel Francis (University of Wales, Swansea (UWS))')
+                    break  #lots of data - just make sure it compiles and runs
+                    
+                self.assertTrue(count > 0)
+            finally:
+                os.remove("%s.py" % name)
         
     def test_complex_datebuilding(self):
         """Loads a pipeline containing a datebuilder with complex inputs
